Reorganize data_sources directory structure
- Move client files to src/data_sources/clients/ directory
- Move constants files to src/data_sources/clients/ directory
- Update import paths in affected files
- Maintain backward compatibility with existing code
